requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==3.2.23
djangorestframework==3.12.4
geographiclib==2.0
geopy==2.4.1
idna==3.10
pip==24.0
pytz==2024.2
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3